factor out projections testcases
